+++ mlx_arsenal/diffusion/__init__.py
@@ -2,14 +2,14 @@
 
 from .samplers import classifier_free_guidance, euler_step
 from .schedulers import (
-    FlowMatchEulerScheduler,
+    FlowMatchEulerDiscreteScheduler,
     dynamic_shift_schedule,
     get_sampling_sigmas,
 )
 from .timestep import TimestepEmbedding, get_timestep_embedding
 
 __all__ = [
-    "FlowMatchEulerScheduler",
+    "FlowMatchEulerDiscreteScheduler",
     "TimestepEmbedding",
     "classifier_free_guidance",
     "dynamic_shift_schedule",